added max load to readings
